Merge in ITB/gitb from development to master

* commit '0a4b25f78bec44a1d4f7f130e6903b124abcf95d': (168 commits)
  Preparation for release 1.22.0
  Added new user guide entries
  Addressed TinyMCE security advisory (as a precaution)
  Bug fixes
  Bug fixes
  Bug fixes
  Bug fixes
  Bug fixes
  Updated ODC to v9.1.0 and resolved CVE-2024-29131/CVE-2024-29133
  Minor bug fixes.
  OpenAPI documentation improvements
  Updated Spring Boot to v3.2.4
  [ITB-1572] Include last update time for test cases and statements
  [ITB-1572] Include last update time for test cases and statements
  [ITB-1572] Include descriptions and system version in conformance snapshots and XML reports
  [ITB-1570] Support setting test configuration via the REST API
  [ITB=1570] Support setting test configuration via the REST API
  Updated the ODC plugin to v9.0.10
  Upgrade to CXF v4.0.4 to resolve (as a precaution) CVE-2024-28752
  [ITB-1578] Define baseline DB scripts for faster setup time on new installations
  ...

+++ gitb-ui/app/managers/TestExecutionManager.scala
@@ -2,7 +2,7 @@ package managers
 
 import actors.SessionManagerActor
 import actors.events.sessions.PrepareTestSessionsEvent
-import akka.actor.{ActorRef, ActorSystem}
+import org.apache.pekko.actor.{ActorRef, ActorSystem}
 import com.gitb.core.{ActorConfiguration, AnyContent, Configuration, ValueEmbeddingEnumeration}
 import exceptions.{AutomationApiException, ErrorCodes, MissingRequiredParameterException}
 import models.Enums.InputMappingMatchType
@@ -464,7 +464,7 @@ class TestExecutionManager @Inject() (testbedClient: managers.TestbedBackendClie
     if (result.isDefined) {
       var reportData: (Option[Path], SessionFolderInfo) = null
       try {
-        reportData = testCaseReportProducer.generateDetailedTestCaseReport(result.get, Some("application/xml"), None)
+        reportData = testCaseReportProducer.generateDetailedTestCaseReport(result.get, Some(Constants.MimeTypeXML), None)
         if (reportData._1.isDefined) {
           Some(Files.readString(reportData._1.get))
         } else {
@@ -473,7 +473,7 @@ class TestExecutionManager @Inject() (testbedClient: managers.TestbedBackendClie
         }
       } catch {
         case e: Exception =>
-          if (reportData._2.archived) {
+          if (reportData != null && reportData._2 != null && reportData._2.archived) {
             FileUtils.deleteQuietly(reportData._2.path.toFile)
           }
           throw e
@@ -503,7 +503,7 @@ class TestExecutionManager @Inject() (testbedClient: managers.TestbedBackendClie
       val report = if (withReports) {
         var reportData: (Option[Path], SessionFolderInfo) = null
         try {
-          reportData = testCaseReportProducer.generateDetailedTestCaseReport(result._1, Some("application/xml"), None)
+          reportData = testCaseReportProducer.generateDetailedTestCaseReport(result._1, Some(Constants.MimeTypeXML), None)
           if (reportData._1.isDefined) {
             Some(Files.readString(reportData._1.get))
           } else {
@@ -512,7 +512,7 @@ class TestExecutionManager @Inject() (testbedClient: managers.TestbedBackendClie
           }
         } catch {
           case e: Exception =>
-            if (reportData._2.archived) {
+            if (reportData != null && reportData._2 != null && reportData._2.archived) {
               FileUtils.deleteQuietly(reportData._2.path.toFile)
             }
             throw e
